feat: 계좌 이체 테스트 코드

+++ bank-app/src/main/java/com/study/bankapp/domain/account/Account.java
@@ -56,7 +56,7 @@ public class Account {
     }
 
     public void checkOwner(Long userId){
-        if(user.getId() != userId){
+        if(user.getId().longValue() != userId.longValue()){
             throw new CustomApiException("계좌 소유자가 아닙니다.");
         }
     }
@@ -67,7 +67,7 @@ public class Account {
 
 
     public void checkSamePassword(Long password) {
-        if(this.password != password){
+        if(this.password.longValue() != password.longValue()){
             throw new CustomApiException("계좌 비밀번호 검증에 실패했습니다.");
         }
     }
